requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.1
cycler==0.10.0
Flask==1.1.2
gunicorn==20.0.4
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
opencv-python==4.2.0.34
Pillow==7.1.1
pyparsing==2.4.7
python-dateutil==2.8.1
requests==2.23.0
selenium==3.141.0
six==1.14.0
urllib3==1.25.9
Werkzeug==1.0.1
